refactor: fin main extraction
