+++ src/runtime/mod.rs
@@ -4,3 +4,4 @@ pub mod prelude;
 pub mod recording;
 pub mod registry;
 pub mod shared;
+pub mod storage;
