+++ src/runtime/prelude.rs
@@ -5,3 +5,4 @@ pub use crate::runtime::gui::*;
 pub use crate::runtime::recording::*;
 pub use crate::runtime::registry::*;
 pub use crate::runtime::shared::*;
+pub use crate::runtime::storage;
